refactor(channels): merge channel-context.ts into types.ts
Eliminates the cross-file import (types.ts → channel-context.ts → types.ts).
LazyChannelContext is 5 lines; isolating it didn't earn its keep.
