+++ src/channels/types.ts
@@ -1,6 +1,5 @@
 import type { Gateway } from "../gateway/index.js";
 import type { Logger } from "../logging/logger.js";
-import type { LazyChannelContext } from "./channel-context.js";
 
 export interface Channel {
   start(deps: ChannelDeps): Promise<void>;
@@ -20,3 +19,14 @@ export interface ChannelActions {
 }
 
 export type ChannelsConfig = Record<string, unknown>;
+
+export interface ChannelContext {
+  getChannelActions(): ChannelActions | undefined;
+}
+
+/** Late-binding so agent tools can be constructed before channels start. */
+export class LazyChannelContext implements ChannelContext {
+  private actions: ChannelActions | undefined;
+  setChannelActions(actions: ChannelActions): void { this.actions = actions; }
+  getChannelActions(): ChannelActions | undefined { return this.actions; }
+}
